refactor: Rename header to navbar
